moved login and signup to Nav component
