+++ src/App.css
@@ -1,0 +1,8 @@
+.App {
+    margin-left: 20px;
+}
+
+input {
+    display: block;
+    margin: 5px 0 5px;
+}
